+++ docs/guide/source/conf.py
@@ -36,6 +36,7 @@ sys.path.append(os.path.abspath("./_extensions/"))
 
 extensions = [
     "sphinx.ext.extlinks",
+    "sphinx_tabs.tabs",
     "docs-rs",
     "gh-example"
 ]